nvif: split out display interface definitions
