+++ drm/nouveau/nv50_display.c
@@ -30,6 +30,12 @@
 #include <drm/drm_dp_helper.h>
 
 #include <nvif/class.h>
+#include <nvif/cl5070.h>
+#include <nvif/cl507a.h>
+#include <nvif/cl507b.h>
+#include <nvif/cl507c.h>
+#include <nvif/cl507d.h>
+#include <nvif/cl507e.h>
 
 #include "nouveau_drm.h"
 #include "nouveau_dma.h"
